Rename institutionAdmin pages to administratorInstitution
