+++ apps/prairielearn/src/ee/pages/administratorInstitutionCourse/administratorInstitutionCourse.sql
