+++ apps/prairielearn/src/ee/pages/administratorInstitutionCourses/administratorInstitutionCourses.html.ts
