+++ apps/prairielearn/src/ee/pages/administratorInstitutionCourses/administratorInstitutionCourses.sql
