+++ apps/prairielearn/src/ee/pages/administratorInstitutionCourses/administratorInstitutionCourses.ts
@@ -2,7 +2,7 @@ import { Router } from 'express';
 import asyncHandler = require('express-async-handler');
 import { loadSqlEquiv, queryRows } from '@prairielearn/postgres';
 import { getInstitution } from '../../lib/institution';
-import { InstitutionAdminCourses } from './institutionAdminCourses.html';
+import { InstitutionAdminCourses } from './administratorInstitutionCourses.html';
 import { CourseSchema } from '../../../lib/db-types';
 
 const sql = loadSqlEquiv(__filename);
