+++ apps/prairielearn/src/ee/pages/administratorInstitutionSaml/administratorInstitutionSaml.ts
@@ -8,7 +8,7 @@ import { z } from 'zod';
 import * as error from '@prairielearn/error';
 import { loadSqlEquiv, queryAsync, runInTransactionAsync } from '@prairielearn/postgres';
 
-import { InstitutionAdminSaml } from './institutionAdminSaml.html';
+import { InstitutionAdminSaml } from './administratorInstitutionSaml.html';
 import {
   getInstitution,
   getInstitutionSamlProvider,
